+++ modules/card-project/card-project.php
@@ -1,6 +1,4 @@
 <?php function telabotanica_module_card_project($data) {
-	global $pug;
-
 	$cover_image_url = bp_attachments_get_attachment('url', array(
 		'object_dir' => 'groups',
 		'item_id' => bp_get_group_id(),
@@ -31,8 +29,46 @@
 	$data = telabotanica_styleguide_data($defaults, $data);
 	$data->modifiers = telabotanica_styleguide_modifiers_array('card-project', $data->modifiers);
 
-	echo $pug->render(__DIR__ . '/card-project.pug', [
-		'data' => $data
-	]);
+	echo '<div class="card-project">';
+		echo '<a class="card-project-link" href="' . $data->permalink . '">';
 
+			echo sprintf(
+				'<div class="card-project-cover" style="background-image: url(%s);">',
+				$data->cover_image_url
+			);
+
+				if ($data->tela) {
+					echo sprintf(
+						'<div class="card-project-tela" title="%s">%s</div>',
+						esc_attr( $data->tela_title ),
+						get_telabotanica_module('icon', ['icon' => 'tela-leaf'])
+					);
+				}
+
+				echo sprintf(
+					'<img src="%s" class="card-project-avatar" />',
+					$data->avatar
+				);
+
+			echo '</div>';
+
+			echo '<div class="card-project-content">';
+				echo '<h2 class="card-project-title"><span>' . $data->name . '</span></h2>';
+				echo '<div class="card-project-description">' . $data->description . '</div>';
+			echo '</div>';
+
+			if ($data->meta) :
+				echo '<div class="card-project-meta">';
+					foreach ($data->meta as $item) :
+						$item = (object) $item;
+						echo '<span class="card-project-meta-item">';
+							the_telabotanica_module('icon', ['icon' => $item->icon]);
+							echo $item->text;
+						echo '</span>';
+					endforeach;
+				echo '</div>';
+			endif;
+
+		echo '</a>';
+	echo '</div>';
 }