+++ modules/icon/icon.php
@@ -1,13 +1,13 @@
 <?php function telabotanica_module_icon($data) {
-	global $pug;
-
-	$defaults = [
-		'color' => false
-	];
-
-	$data = telabotanica_styleguide_data($defaults, $data);
-
-	echo $pug->render(__DIR__ . '/icon.pug', [
-		'data' => $data
-	]);
+  if ( isset($data->color) ) {
+    $data->color = 'icon-color-' . $data->color;
+  } else {
+    $data->color = '';
+  }
+  echo sprintf(
+    '<svg aria-hidden="true" role="img" class="icon icon-%s %s"><use xlink:href="#icon-%s"></use></svg>',
+    $data->icon,
+    $data->color,
+    $data->icon
+  );
 }